+++ R/matchEnvData.R
@@ -19,6 +19,10 @@
 #'   will attempt to download a single nc file covering the entire range of your data.
 #'   If your data spans a large amount of time and space this can be problematic.
 #' @param progress logical flag to show progress bar
+#' @param depth depth values (meters) to use for matching, overrides any \code{Depth} column
+#'   in the data or can be used to specify desired depth range when not present in data.
+#'   Variables will be summarised over the range of these depth values. \code{NULL}
+#'   uses all available depth values
 #' @param \dots other parameters to pass to \link{ncToData}
 #'
 #' @return original dataframe with three attached columns for each variable in the netcdf
@@ -55,7 +59,7 @@
 #'
 setGeneric('matchEnvData',
            function(data, nc=NULL, var=NULL, buffer=c(0,0,0), FUN = c(mean),
-                    fileName = NULL, progress=TRUE, ...) standardGeneric('matchEnvData')
+                    fileName = NULL, progress=TRUE, depth=0, ...) standardGeneric('matchEnvData')
 )
 
 #' @rdname matchEnvData
@@ -66,7 +70,7 @@ setGeneric('matchEnvData',
 #'
 setMethod('matchEnvData', 'data.frame',
           function(data, nc=NULL, var=NULL, buffer=c(0,0,0), FUN = c(mean),
-                   fileName = NULL, progress=TRUE, ...) {
+                   fileName = NULL, progress=TRUE, depth=0, ...) {
               # First just get an edinfo
               if(is.null(nc)) {
                   nc <- browseEdinfo(var=var)
@@ -90,7 +94,7 @@ setMethod('matchEnvData', 'data.frame',
               # if pointing to an ncfile, just do that
               if(is.character(nc) &&
                  file.exists(nc)) {
-                  return(ncToData(data=data, nc=nc, buffer=buffer, FUN=FUN, progress=progress, ...))
+                  return(ncToData(data=data, nc=nc, buffer=buffer, FUN=FUN, progress=progress, depth=depth, ...))
               }
 
               if(!inherits(nc, 'edinfo')) {
@@ -119,7 +123,7 @@ setMethod('matchEnvData', 'data.frame',
                       if(hys[i] == -1) next
                       thisHy <- nc$list[[hys[i]]]
                       thisHy$varSelect <- nc$varSelect
-                      result[[i]] <- matchEnvData(data[whichHy == hys[i], ], nc=thisHy, var, buffer, FUN, fileName, progress, ...)
+                      result[[i]] <- matchEnvData(data[whichHy == hys[i], ], nc=thisHy, var, buffer, FUN, fileName, progress, depth, ...)
                       fixer <- c(fixer, which(whichHy == hys[i]))
                   }
                   fixer <- sort(fixer, index.return=TRUE)$ix
@@ -161,7 +165,7 @@ setMethod('matchEnvData', 'data.frame',
                   }
                   result <- vector('list', length = nrow(data))
                   for(i in seq_along(result)) {
-                      result[[i]] <- ncToData(data=data[i, ], nc=planFiles[[plan[i]]], buffer=buffer, FUN=FUN, progress=FALSE, ...)
+                      result[[i]] <- ncToData(data=data[i, ], nc=planFiles[[plan[i]]], buffer=buffer, FUN=FUN, progress=FALSE, depth=depth, ...)
                   }
                   if(progress) {
                       cat('\n')
@@ -189,11 +193,11 @@ setMethod('matchEnvData', 'data.frame',
                   colnames(data) <- oldNames
                   dataLeft <- data[left, ]
                   dataRight <- data[!left, ]
-                  return(bind_rows(ncToData(data=dataLeft, nc=ncData[1], buffer=buffer, FUN=FUN, progress=progress, ...),
-                                   ncToData(data=dataRight, nc=ncData[2], buffer=buffer, FUN=FUN, progress=progress, ...))
+                  return(bind_rows(ncToData(data=dataLeft, nc=ncData[1], buffer=buffer, FUN=FUN, progress=progress, depth=depth, ...),
+                                   ncToData(data=dataRight, nc=ncData[2], buffer=buffer, FUN=FUN, progress=progress, depth=depth, ...))
                   )
               }
-              return(ncToData(data=data, nc=ncData, buffer=buffer, FUN=FUN, progress=progress, ...))
+              return(ncToData(data=data, nc=ncData, buffer=buffer, FUN=FUN, progress=progress, depth=depth, ...))
           })
 
 fillNA <- function(x, ix) {